backend/requirements.txt
Flask>=2.3.0
Flask-CORS>=4.0.0
python-dotenv>=1.0.0
requests>=2.28.0
supabase>=1.0.0,<2.0.0
python-telegram-bot>=20.0
